@@ -46,7 +46,7 @@
     "sync-schema": "curl https://raw.githubusercontent.com/artsy/metaphysics/master/_schemaV2.graphql -o data/schema.graphql; yarn prettier --write --parser graphql data/schema.graphql",
     "sync-cities": "curl https://raw.githubusercontent.com/artsy/metaphysics/master/src/schema/city/cityDataSortedByDisplayPreference.json -o data/cityDataSortedByDisplayPreference.json",
     "version": "yarn install && yarn bundle && pushd Example && bundle install && bundle exec pod update Emission --no-repo-update && popd && git add Pod/Assets Example/Podfile.lock",
-    "postinstall": "prettier --write package.json",
+    "postinstall": "prettier --write package.json && node verify-npm-podspecs.js",
     "prestorybook": "rnstl --silent",
     "prepare": "patch-package"
   },
@@ -76,25 +76,28 @@
   },
   "dependencies": {
     "@artsy/palette": "5.1.6",
-    "@mapbox/react-native-mapbox-gl": "l2succes/react-native-mapbox-gl#fix-gesture-recognizer",
-    "@react-native-community/cameraroll": "1.0.5",
+    "@mapbox/react-native-mapbox-gl": "^6.1.3",
+    "@react-native-community/cameraroll": "^1.3.0",
+    "@react-native-community/geolocation": "^2.0.2",
+    "@react-native-community/netinfo": "^4.6.1",
     "lodash": "4.17.15",
     "moment": "2.24.0",
     "moment-timezone": "0.5.25",
     "prop-types": "15.7.2",
     "query-string": "4.3.4",
-    "react": "16.8.3",
-    "react-native": "0.59.2",
+    "react": "16.9.0",
+    "react-native": "0.61.4",
     "react-native-hyperlink": "0.0.11",
     "react-native-navigator-ios": "https://github.com/ashfurrow/react-native-navigator-ios#license_podspec",
     "react-native-parallax-scroll-view": "orta/react-native-parallax-scroll-view",
     "react-native-reanimated": "1.4.0",
-    "react-native-scrollable-tab-view": "0.9.0",
+    "react-native-scrollable-tab-view": "1.0.0",
     "react-native-sentry": "0.30.3",
-    "react-native-svg": "9.4.0",
+    "react-native-svg": "^9.13.3",
+    "react-native-webview": "^7.5.1",
     "react-relay": "6.0.0",
     "react-spring": "8.0.23",
-    "react-tracking": "7.0.1",
+    "react-tracking": "^7.0.1",
     "relay-runtime": "6.0.0",
     "remove-markdown": "0.1.0",
     "simple-markdown": "0.4.4",
@@ -112,6 +115,7 @@
     "@babel/runtime": "7.4.2",
     "@octokit/rest": "16.34.0",
     "@omakase/cli": "0.0.10",
+    "@react-native-community/cli-platform-ios": "^3.0.0",
     "@storybook/react-native": "3.3.11",
     "@types/chalk": "2.2.0",
     "@types/dedent": "0.7.0",
