@@ -44,25 +44,6 @@ fragment WorksForYou_query_1G22uz on Query {
     }
     id
   }
-  selectedArtist: artist(id: "") {
-    slug
-    href
-    name
-    image {
-      resized(height: 80, width: 80) {
-        url
-      }
-    }
-    artworks: artworksConnection(sort: PUBLISHED_AT_DESC, first: 6) {
-      edges {
-        node {
-          ...GenericGrid_artworks
-          id
-        }
-      }
-    }
-    id
-  }
 }
 
 fragment Notification_notification on FollowedArtistsArtworksGroup {
@@ -146,12 +127,7 @@ var v0 = [
     "defaultValue": null
   }
 ],
-v1 = {
-  "kind": "Literal",
-  "name": "sort",
-  "value": "PUBLISHED_AT_DESC"
-},
-v2 = [
+v1 = [
   {
     "kind": "Variable",
     "name": "after",
@@ -162,232 +138,32 @@ v2 = [
     "name": "first",
     "variableName": "count"
   },
-  (v1/*: any*/)
+  {
+    "kind": "Literal",
+    "name": "sort",
+    "value": "PUBLISHED_AT_DESC"
+  }
 ],
-v3 = {
+v2 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "id",
   "args": null,
   "storageKey": null
 },
-v4 = [
-  {
-    "kind": "Literal",
-    "name": "shallow",
-    "value": true
-  }
-],
-v5 = {
+v3 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "href",
   "args": null,
   "storageKey": null
 },
-v6 = {
+v4 = {
   "kind": "ScalarField",
   "alias": null,
   "name": "name",
   "args": null,
   "storageKey": null
-},
-v7 = {
-  "kind": "LinkedField",
-  "alias": null,
-  "name": "image",
-  "storageKey": null,
-  "args": null,
-  "concreteType": "Image",
-  "plural": false,
-  "selections": [
-    {
-      "kind": "ScalarField",
-      "alias": "aspect_ratio",
-      "name": "aspectRatio",
-      "args": null,
-      "storageKey": null
-    },
-    {
-      "kind": "ScalarField",
-      "alias": null,
-      "name": "url",
-      "args": [
-        {
-          "kind": "Literal",
-          "name": "version",
-          "value": "large"
-        }
-      ],
-      "storageKey": "url(version:\"large\")"
-    }
-  ]
-},
-v8 = {
-  "kind": "ScalarField",
-  "alias": null,
-  "name": "title",
-  "args": null,
-  "storageKey": null
-},
-v9 = {
-  "kind": "ScalarField",
-  "alias": null,
-  "name": "date",
-  "args": null,
-  "storageKey": null
-},
-v10 = {
-  "kind": "ScalarField",
-  "alias": "sale_message",
-  "name": "saleMessage",
-  "args": null,
-  "storageKey": null
-},
-v11 = {
-  "kind": "ScalarField",
-  "alias": "is_biddable",
-  "name": "isBiddable",
-  "args": null,
-  "storageKey": null
-},
-v12 = {
-  "kind": "ScalarField",
-  "alias": "is_acquireable",
-  "name": "isAcquireable",
-  "args": null,
-  "storageKey": null
-},
-v13 = {
-  "kind": "ScalarField",
-  "alias": "is_offerable",
-  "name": "isOfferable",
-  "args": null,
-  "storageKey": null
-},
-v14 = {
-  "kind": "ScalarField",
-  "alias": null,
-  "name": "slug",
-  "args": null,
-  "storageKey": null
-},
-v15 = {
-  "kind": "LinkedField",
-  "alias": null,
-  "name": "sale",
-  "storageKey": null,
-  "args": null,
-  "concreteType": "Sale",
-  "plural": false,
-  "selections": [
-    {
-      "kind": "ScalarField",
-      "alias": "is_auction",
-      "name": "isAuction",
-      "args": null,
-      "storageKey": null
-    },
-    {
-      "kind": "ScalarField",
-      "alias": "is_closed",
-      "name": "isClosed",
-      "args": null,
-      "storageKey": null
-    },
-    {
-      "kind": "ScalarField",
-      "alias": "display_timely_at",
-      "name": "displayTimelyAt",
-      "args": null,
-      "storageKey": null
-    },
-    (v3/*: any*/)
-  ]
-},
-v16 = {
-  "kind": "LinkedField",
-  "alias": "sale_artwork",
-  "name": "saleArtwork",
-  "storageKey": null,
-  "args": null,
-  "concreteType": "SaleArtwork",
-  "plural": false,
-  "selections": [
-    {
-      "kind": "LinkedField",
-      "alias": "current_bid",
-      "name": "currentBid",
-      "storageKey": null,
-      "args": null,
-      "concreteType": "SaleArtworkCurrentBid",
-      "plural": false,
-      "selections": [
-        {
-          "kind": "ScalarField",
-          "alias": null,
-          "name": "display",
-          "args": null,
-          "storageKey": null
-        }
-      ]
-    },
-    (v3/*: any*/)
-  ]
-},
-v17 = [
-  (v6/*: any*/),
-  (v3/*: any*/)
-],
-v18 = {
-  "kind": "LinkedField",
-  "alias": null,
-  "name": "partner",
-  "storageKey": null,
-  "args": null,
-  "concreteType": "Partner",
-  "plural": false,
-  "selections": (v17/*: any*/)
-},
-v19 = {
-  "kind": "LinkedField",
-  "alias": null,
-  "name": "image",
-  "storageKey": null,
-  "args": null,
-  "concreteType": "Image",
-  "plural": false,
-  "selections": [
-    {
-      "kind": "LinkedField",
-      "alias": null,
-      "name": "resized",
-      "storageKey": "resized(height:80,width:80)",
-      "args": [
-        {
-          "kind": "Literal",
-          "name": "height",
-          "value": 80
-        },
-        {
-          "kind": "Literal",
-          "name": "width",
-          "value": 80
-        }
-      ],
-      "concreteType": "ResizedImageUrl",
-      "plural": false,
-      "selections": [
-        {
-          "kind": "ScalarField",
-          "alias": null,
-          "name": "url",
-          "args": null,
-          "storageKey": null
-        }
-      ]
-    }
-  ]
 };
 return {
   "kind": "Request",
@@ -444,7 +220,7 @@ return {
                 "alias": "notifications",
                 "name": "bundledArtworksByArtistConnection",
                 "storageKey": null,
-                "args": (v2/*: any*/),
+                "args": (v1/*: any*/),
                 "concreteType": "FollowedArtistsArtworksGroupConnection",
                 "plural": false,
                 "selections": [
@@ -491,7 +267,7 @@ return {
                         "concreteType": "FollowedArtistsArtworksGroup",
                         "plural": false,
                         "selections": [
-                          (v3/*: any*/),
+                          (v2/*: any*/),
                           {
                             "kind": "ScalarField",
                             "alias": null,
@@ -544,35 +320,225 @@ return {
                                         "alias": null,
                                         "name": "artists",
                                         "storageKey": "artists(shallow:true)",
-                                        "args": (v4/*: any*/),
+                                        "args": [
+                                          {
+                                            "kind": "Literal",
+                                            "name": "shallow",
+                                            "value": true
+                                          }
+                                        ],
                                         "concreteType": "Artist",
                                         "plural": true,
                                         "selections": [
-                                          (v5/*: any*/),
                                           (v3/*: any*/),
-                                          (v6/*: any*/)
+                                          (v2/*: any*/),
+                                          (v4/*: any*/)
                                         ]
                                       },
-                                      (v3/*: any*/),
-                                      (v7/*: any*/),
-                                      (v8/*: any*/),
-                                      (v9/*: any*/),
-                                      (v10/*: any*/),
-                                      (v11/*: any*/),
-                                      (v12/*: any*/),
-                                      (v13/*: any*/),
-                                      (v14/*: any*/),
-                                      (v15/*: any*/),
-                                      (v16/*: any*/),
-                                      (v18/*: any*/),
-                                      (v5/*: any*/)
+                                      (v2/*: any*/),
+                                      {
+                                        "kind": "LinkedField",
+                                        "alias": null,
+                                        "name": "image",
+                                        "storageKey": null,
+                                        "args": null,
+                                        "concreteType": "Image",
+                                        "plural": false,
+                                        "selections": [
+                                          {
+                                            "kind": "ScalarField",
+                                            "alias": "aspect_ratio",
+                                            "name": "aspectRatio",
+                                            "args": null,
+                                            "storageKey": null
+                                          },
+                                          {
+                                            "kind": "ScalarField",
+                                            "alias": null,
+                                            "name": "url",
+                                            "args": [
+                                              {
+                                                "kind": "Literal",
+                                                "name": "version",
+                                                "value": "large"
+                                              }
+                                            ],
+                                            "storageKey": "url(version:\"large\")"
+                                          }
+                                        ]
+                                      },
+                                      {
+                                        "kind": "ScalarField",
+                                        "alias": null,
+                                        "name": "title",
+                                        "args": null,
+                                        "storageKey": null
+                                      },
+                                      {
+                                        "kind": "ScalarField",
+                                        "alias": null,
+                                        "name": "date",
+                                        "args": null,
+                                        "storageKey": null
+                                      },
+                                      {
+                                        "kind": "ScalarField",
+                                        "alias": "sale_message",
+                                        "name": "saleMessage",
+                                        "args": null,
+                                        "storageKey": null
+                                      },
+                                      {
+                                        "kind": "ScalarField",
+                                        "alias": "is_biddable",
+                                        "name": "isBiddable",
+                                        "args": null,
+                                        "storageKey": null
+                                      },
+                                      {
+                                        "kind": "ScalarField",
+                                        "alias": "is_acquireable",
+                                        "name": "isAcquireable",
+                                        "args": null,
+                                        "storageKey": null
+                                      },
+                                      {
+                                        "kind": "ScalarField",
+                                        "alias": "is_offerable",
+                                        "name": "isOfferable",
+                                        "args": null,
+                                        "storageKey": null
+                                      },
+                                      {
+                                        "kind": "ScalarField",
+                                        "alias": null,
+                                        "name": "slug",
+                                        "args": null,
+                                        "storageKey": null
+                                      },
+                                      {
+                                        "kind": "LinkedField",
+                                        "alias": null,
+                                        "name": "sale",
+                                        "storageKey": null,
+                                        "args": null,
+                                        "concreteType": "Sale",
+                                        "plural": false,
+                                        "selections": [
+                                          {
+                                            "kind": "ScalarField",
+                                            "alias": "is_auction",
+                                            "name": "isAuction",
+                                            "args": null,
+                                            "storageKey": null
+                                          },
+                                          {
+                                            "kind": "ScalarField",
+                                            "alias": "is_closed",
+                                            "name": "isClosed",
+                                            "args": null,
+                                            "storageKey": null
+                                          },
+                                          {
+                                            "kind": "ScalarField",
+                                            "alias": "display_timely_at",
+                                            "name": "displayTimelyAt",
+                                            "args": null,
+                                            "storageKey": null
+                                          },
+                                          (v2/*: any*/)
+                                        ]
+                                      },
+                                      {
+                                        "kind": "LinkedField",
+                                        "alias": "sale_artwork",
+                                        "name": "saleArtwork",
+                                        "storageKey": null,
+                                        "args": null,
+                                        "concreteType": "SaleArtwork",
+                                        "plural": false,
+                                        "selections": [
+                                          {
+                                            "kind": "LinkedField",
+                                            "alias": "current_bid",
+                                            "name": "currentBid",
+                                            "storageKey": null,
+                                            "args": null,
+                                            "concreteType": "SaleArtworkCurrentBid",
+                                            "plural": false,
+                                            "selections": [
+                                              {
+                                                "kind": "ScalarField",
+                                                "alias": null,
+                                                "name": "display",
+                                                "args": null,
+                                                "storageKey": null
+                                              }
+                                            ]
+                                          },
+                                          (v2/*: any*/)
+                                        ]
+                                      },
+                                      {
+                                        "kind": "LinkedField",
+                                        "alias": null,
+                                        "name": "partner",
+                                        "storageKey": null,
+                                        "args": null,
+                                        "concreteType": "Partner",
+                                        "plural": false,
+                                        "selections": [
+                                          (v4/*: any*/),
+                                          (v2/*: any*/)
+                                        ]
+                                      },
+                                      (v3/*: any*/)
                                     ]
                                   }
                                 ]
                               }
                             ]
                           },
-                          (v19/*: any*/),
+                          {
+                            "kind": "LinkedField",
+                            "alias": null,
+                            "name": "image",
+                            "storageKey": null,
+                            "args": null,
+                            "concreteType": "Image",
+                            "plural": false,
+                            "selections": [
+                              {
+                                "kind": "LinkedField",
+                                "alias": null,
+                                "name": "resized",
+                                "storageKey": "resized(height:80,width:80)",
+                                "args": [
+                                  {
+                                    "kind": "Literal",
+                                    "name": "height",
+                                    "value": 80
+                                  },
+                                  {
+                                    "kind": "Literal",
+                                    "name": "width",
+                                    "value": 80
+                                  }
+                                ],
+                                "concreteType": "ResizedImageUrl",
+                                "plural": false,
+                                "selections": [
+                                  {
+                                    "kind": "ScalarField",
+                                    "alias": null,
+                                    "name": "url",
+                                    "args": null,
+                                    "storageKey": null
+                                  }
+                                ]
+                              }
+                            ]
+                          },
                           {
                             "kind": "ScalarField",
                             "alias": null,
@@ -597,7 +563,7 @@ return {
                 "kind": "LinkedHandle",
                 "alias": "notifications",
                 "name": "bundledArtworksByArtistConnection",
-                "args": (v2/*: any*/),
+                "args": (v1/*: any*/),
                 "handle": "connection",
                 "key": "WorksForYou_notifications",
                 "filters": [
@@ -606,92 +572,7 @@ return {
               }
             ]
           },
-          (v3/*: any*/)
-        ]
-      },
-      {
-        "kind": "LinkedField",
-        "alias": "selectedArtist",
-        "name": "artist",
-        "storageKey": "artist(id:\"\")",
-        "args": [
-          {
-            "kind": "Literal",
-            "name": "id",
-            "value": ""
-          }
-        ],
-        "concreteType": "Artist",
-        "plural": false,
-        "selections": [
-          (v14/*: any*/),
-          (v5/*: any*/),
-          (v6/*: any*/),
-          (v19/*: any*/),
-          {
-            "kind": "LinkedField",
-            "alias": "artworks",
-            "name": "artworksConnection",
-            "storageKey": "artworksConnection(first:6,sort:\"PUBLISHED_AT_DESC\")",
-            "args": [
-              {
-                "kind": "Literal",
-                "name": "first",
-                "value": 6
-              },
-              (v1/*: any*/)
-            ],
-            "concreteType": "ArtworkConnection",
-            "plural": false,
-            "selections": [
-              {
-                "kind": "LinkedField",
-                "alias": null,
-                "name": "edges",
-                "storageKey": null,
-                "args": null,
-                "concreteType": "ArtworkEdge",
-                "plural": true,
-                "selections": [
-                  {
-                    "kind": "LinkedField",
-                    "alias": null,
-                    "name": "node",
-                    "storageKey": null,
-                    "args": null,
-                    "concreteType": "Artwork",
-                    "plural": false,
-                    "selections": [
-                      (v3/*: any*/),
-                      (v7/*: any*/),
-                      (v8/*: any*/),
-                      (v9/*: any*/),
-                      (v10/*: any*/),
-                      (v11/*: any*/),
-                      (v12/*: any*/),
-                      (v13/*: any*/),
-                      (v14/*: any*/),
-                      (v15/*: any*/),
-                      (v16/*: any*/),
-                      {
-                        "kind": "LinkedField",
-                        "alias": null,
-                        "name": "artists",
-                        "storageKey": "artists(shallow:true)",
-                        "args": (v4/*: any*/),
-                        "concreteType": "Artist",
-                        "plural": true,
-                        "selections": (v17/*: any*/)
-                      },
-                      (v18/*: any*/),
-                      (v5/*: any*/)
-                    ]
-                  }
-                ]
-              }
-            ]
-          },
-          (v3/*: any*/)
+          (v2/*: any*/)
         ]
       }
     ]
@@ -699,7 +580,7 @@ return {
   "params": {
     "operationKind": "query",
     "name": "WorksForYouQuery",
-    "id": "cc26a9f5919786a4fa9cdc103f814269",
+    "id": "233fefb662c9aabcdda8d976ab43f755",
     "text": null,
     "metadata": {}
   }
