@@ -1,8 +1,15 @@
-import { Box, Flex, Theme } from "@artsy/palette"
+import { Box, Flex, Separator, Theme } from "@artsy/palette"
+import { WorksForYou_query } from "__generated__/WorksForYou_query.graphql"
+import Spinner from "lib/Components/Spinner"
+import { ZeroState } from "lib/Components/States/ZeroState"
+import Notification from "lib/Components/WorksForYou/Notification"
+import colors from "lib/data/colors"
+import { PAGE_SIZE } from "lib/data/constants"
+import { get } from "lib/utils/get"
+import { isCloseToBottom } from "lib/utils/isCloseToBottom"
 import React from "react"
 import {
-  ListView,
-  ListViewDataSource,
+  FlatList,
   NativeModules,
   RefreshControl,
   ScrollView,
@@ -12,17 +19,7 @@ import {
   ViewStyle,
 } from "react-native"
 import { createPaginationContainer, graphql, RelayPaginationProp } from "react-relay"
-
 import Events from "../NativeModules/Events"
-
-import { PAGE_SIZE } from "lib/data/constants"
-
-import { WorksForYou_query } from "__generated__/WorksForYou_query.graphql"
-import Spinner from "lib/Components/Spinner"
-import { ZeroState } from "lib/Components/States/ZeroState"
-import Notification from "lib/Components/WorksForYou/Notification"
-import colors from "lib/data/colors"
-import { isCloseToBottom } from "lib/utils/isCloseToBottom"
 
 interface Props {
   relay: RelayPaginationProp
@@ -30,28 +27,17 @@ interface Props {
 }
 
 interface State {
-  dataSource: ListViewDataSource | null
   isRefreshing: boolean
   loadingContent: boolean
 }
 
 export class WorksForYou extends React.Component<Props, State> {
-  // TODO: This `| any` is a hack workaround to a typing bug in https://github.com/artsy/emission/pull/504/
-  scrollView?: ScrollView | any
   currentScrollOffset?: number = 0
 
   constructor(props: Props) {
     super(props)
 
-    const notifications: object[] = this.props.query.me.followsAndSaves.notifications.edges.map(edge => edge.node)
-    if (this.props.query.selectedArtist) {
-      notifications.unshift(this.formattedSpecialNotification())
-    }
-
-    const dataSource = createDataSource(notifications)
-
     this.state = {
-      dataSource,
       isRefreshing: false,
       loadingContent: false,
     }
@@ -71,24 +57,6 @@ export class WorksForYou extends React.Component<Props, State> {
     })
   }
 
-  formattedSpecialNotification() {
-    const artist = this.props.query.selectedArtist
-
-    return {
-      // This is just some unique ID, don’t rely on MP being able to retrieve a notification by this ID.
-      id: `notification-${artist.slug}`,
-      message: artist.artworks.edges.length + (artist.artworks.edges.length > 1 ? " Works Added" : " Work Added"),
-      artists: artist.name,
-      artworks: artist.artworks,
-      image: {
-        resized: {
-          url: artist.image.resized.url,
-        },
-      },
-      artistHref: artist.href,
-    }
-  }
-
   fetchNextPage = () => {
     if (!this.props.relay.hasMore() || this.props.relay.isLoading()) {
       return
@@ -100,21 +68,7 @@ export class WorksForYou extends React.Component<Props, State> {
           console.error("WorksForYou.tsx", error.message)
         }
 
-        const notifications: object[] = this.props.query.me.followsAndSaves.notifications.edges.map(edge => edge.node)
-
-        // Make sure we maintain the special notification if it exists
-        if (this.props.query.selectedArtist) {
-          notifications.unshift(this.formattedSpecialNotification())
-        }
-
-        this.setState(
-          {
-            dataSource: this.state.dataSource.cloneWithRows(notifications),
-          },
-          () => {
-            this.setState({ loadingContent: false })
-          }
-        )
+        this.setState({ loadingContent: false })
       })
     })
   }
@@ -126,14 +80,13 @@ export class WorksForYou extends React.Component<Props, State> {
         // FIXME: Handle error
         console.error("WorksForYou.tsx #handleRefresh", error.message)
       }
-      // update data source from scratch
-      const notifications: object[] = this.props.query.me.followsAndSaves.notifications.edges.map(edge => edge.node)
-      this.setState({ isRefreshing: false, dataSource: createDataSource(notifications) })
+      this.setState({ isRefreshing: false })
     })
   }
 
   render() {
-    const hasNotifications = this.state.dataSource
+    const notifications = get(this.props, props => props.query.me.followsAndSaves.notifications.edges)
+    const hasNotifications = notifications.length
 
     /* If showing the empty state, the ScrollView should have a {flex: 1} style so it can expand to fit the screen.
        otherwise, it should not use any flex growth.
@@ -143,7 +96,6 @@ export class WorksForYou extends React.Component<Props, State> {
         contentContainerStyle={hasNotifications ? {} : styles.container}
         onScroll={isCloseToBottom(this.fetchNextPage)}
         scrollEventThrottle={100}
-        ref={scrollView => (this.scrollView = scrollView)}
         refreshControl={<RefreshControl refreshing={this.state.isRefreshing} onRefresh={this.handleRefresh} />}
       >
         <View style={{ flex: 1 }}>{hasNotifications ? this.renderNotifications() : this.renderEmptyState()}</View>
@@ -153,16 +105,23 @@ export class WorksForYou extends React.Component<Props, State> {
 
   renderNotifications() {
     const { loadingContent } = this.state
+    const notifications = get(this.props, props => props.query.me.followsAndSaves.notifications.edges)
 
     return (
       <Theme>
-        <ListView
-          dataSource={this.state.dataSource}
-          renderRow={data => <Notification notification={data} />}
-          renderSeparator={(sectionID, rowID) =>
-            <View key={`${sectionID}-${rowID}`} style={styles.separator} /> as React.ReactElement<{}>
-          }
-          renderFooter={() => (
+        <FlatList
+          data={notifications}
+          keyExtractor={item => item.node.id}
+          scrollEnabled={false}
+          renderItem={data => {
+            return <Notification notification={data.item.node} />
+          }}
+          ItemSeparatorComponent={() => (
+            <Box px={2}>
+              <Separator />
+            </Box>
+          )}
+          ListFooterComponent={() => (
             <>
               {loadingContent && (
                 <Box p={2} style={{ height: 50 }}>
@@ -173,7 +132,6 @@ export class WorksForYou extends React.Component<Props, State> {
               )}
             </>
           )}
-          scrollEnabled={false}
         />
       </Theme>
     )
@@ -217,7 +175,6 @@ const WorksForYouContainer = createPaginationContainer(
         @argumentDefinitions(
           count: { type: "Int", defaultValue: 10 }
           cursor: { type: "String" }
-          selectedArtist: { type: "String!", defaultValue: "" }
           sort: { type: "ArtworkSorts" }
         ) {
         me {
@@ -233,23 +190,6 @@ const WorksForYouContainer = createPaginationContainer(
                   id
                   ...Notification_notification
                 }
-              }
-            }
-          }
-        }
-        selectedArtist: artist(id: $selectedArtist) {
-          slug
-          href
-          name
-          image {
-            resized(height: 80, width: 80) {
-              url
-            }
-          }
-          artworks: artworksConnection(sort: PUBLISHED_AT_DESC, first: 6) {
-            edges {
-              node {
-                ...GenericGrid_artworks
               }
             }
           }
@@ -284,9 +224,5 @@ const WorksForYouContainer = createPaginationContainer(
     `,
   }
 )
-
-const createDataSource = (notifications: object[]) =>
-  notifications.length &&
-  new ListView.DataSource({ rowHasChanged: (row1, row2) => row1 !== row2 }).cloneWithRows(notifications)
 
 export default WorksForYouContainer
